settings.cfg, colors
Previously, the representation of the refined match used
two tints of gray for test and reference motif.  Neither
was this easing the discern of more similar model geometries,
nor was this consequent to the color sheme introduced in
the representation prior to refine the superposition.

Now said more distinctive color scheme is used again.