alog.py reformat details matching process
In the previous layout, the entry on lines #59 and following
(aRMSD_logfile.out) about 'The highest deviations were between
the pairs...' was table-like organized.  The two corresponding
'columns' were moved a bit to the right, so that the one about
the pairs is centered just below 'between', and the numeric
output just below '[Angstrom]'.