update counter and license
The previous changes to the project seem to justify
an increment of the version at the level of bugfix.
The counter was updated accordingly.  The license was
updated by date and where changes made by mine, my
name was added.

Since the working defined by aplot.py was not altered
by mine, I revoked my name from the copyright line.