alog.py, reformatting output Kabsch test
Among the similarities provided (cosine, GARD) after refined
alignment of the two model data, the result by the Kabsch test
already was present.  But the line about the RMSD benefits
from an explicit statement that /this one/ is about the resiudal
error of the Kabsch test; which was edited accordingly.