aplot.py reverting \AA
Reversion of the of commit by October 15th, 3:00 pm, hash
78d194eaf8d9b3f6c241e19e555e3ea842d11bdc; which was changing
the formatting of \AA in the statistics plots.  A glimpse of
using aRMSD under Linux Xubuntu 16.04 LTS (point release 16.04.5)
for the first time suggests the fonts used in the mathematical
mode may, but not necessarily do contain a sans serif upper-case
glyph \AA matching the other around.