alog.py format issues
Linting the code to a degree that /most likely/ the permanent
log's layout does not break if the inner tabulators remaining
are interpreted as four spaces (a la Python) or lengthier ones
(eight?).

Changing the order of the results: now, Kabsch' data preceed
the other similarities.  Parameters and results of the GARD
calculation remain in neighborhood to each other and the
additional literature reference.